backend/requirements2.txt
aiohttp==3.7.4.post0
aiohttp-sse-client==0.2.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
cryptography==3.4.8
dj-rest-auth==2.2.1
Django==3.2.8
django-cors-headers==3.10.0
django-environ==0.7.0
django-model-utils==4.1.1
django-polaris==2.0.5
django-rest-authtoken==2.1.4
djangorestframework==3.12.4
frozenlist==1.2.0
gunicorn==20.1.0
idna==3.2
mnemonic==0.20
multidict==5.1.0
mysqlclient==2.0.3
Pillow==8.4.0
pycparser==2.20
PyJWT==2.2.0
PyNaCl==1.4.0
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
stellar-base-sseclient==0.0.21
stellar-sdk==5.0.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
whitenoise==5.3.0
yarl==1.6.3
